requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
langchain==0.3.14
langchain-groq==0.2.1
